Refactor code to MVC
