+++ 4-natours/starter/controllers/tourController.js
@@ -1,18 +1,4 @@
-const fs = require('fs');
-
-const tours = JSON.parse(
-  fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`)
-);
-
-exports.checkId = (req, res, next) => {
-  if (+req.params.id > tours.length) {
-    return res.status(404).json({
-      status: 'fail',
-      message: 'Invalid ID',
-    });
-  }
-  next();
-};
+const tourModel = require('./../models/tourModel');
 
 exports.checkBody = (req, res, next) => {
   if (!req.body.name || !req.body.price) {
@@ -29,45 +15,28 @@ exports.getAllTours = (req, res) => {
   res.status(200).json({
     status: 'success',
     requestedAt: req.requestTime,
-    results: tours.length,
-    data: { tours },
+    // results: tours.length,
+    // data: { tours },
   });
 };
 
 exports.getTour = (req, res) => {
-  const id = +req.params.id;
-  const tour = tours.find((el) => el.id === id);
-
-  //console.log(req.params);
-  res.status(200).json({
-    status: 'success',
-    data: { tour },
-  });
+  //const id = +req.params.id;
+  // const tour = tours.find((el) => el.id === id);
+  // //console.log(req.params);
+  // res.status(200).json({
+  //   status: 'success',
+  //   data: { tour },
+  // });
 };
 
 exports.createTour = (req, res) => {
-  //console.log(req.body);
-
-  const newId = tours[tours.length - 1].id + 1;
-  const newTour = Object.assign(req.body, { id: newId });
-
-  tours.push(newTour);
-  fs.writeFile(
-    `${__dirname}/dev-data/data/tours-simple.json`,
-    JSON.stringify(tours),
-    () => {
-      res.status(201).json({
-        status: 'success',
-        data: {
-          tour: newTour,
-        },
-      });
-    }
-  );
-  // res.status(204).json({
-  //   status: 'success',
-  //   data: null,
-  // });
+  res.status(201).json({
+    status: 'success',
+    // data: {
+    //   tour: newTour,
+    // },
+  });
 };
 
 exports.updateTour = (req, res) => {
